+++ metadata.php
@@ -284,7 +284,8 @@ $aModule = [
             'name' => 'oscPayPalStandardCaptureStrategy',
             'type' => 'select',
             'value' => 'directly',
-            'constraints' => 'directly|delivery|manually'
+            'constraints' => 'directly|delivery|manually',
+            'group' => null
         ],
         [
             'name' => 'oscPayPalSetupFeeFailureAction',
@@ -350,7 +351,8 @@ $aModule = [
         [
             'name' => 'oscPayPalBannersProductDetailsPageSelector',
             'type' => 'str',
-            'value' => '.breadcrumb-wrapper > .container-xxl'
+            'value' => '.breadcrumb-wrapper > .container-xxl',
+            'group' => null
         ],
         [
             'name' => 'oscPayPalBannersCheckoutPage',